+++ trunk/source/menu/menu_config3.cpp
@@ -41,8 +41,8 @@ void CMenu::_hideConfig3(bool instant)
 	m_btnMgr.hide(m_config3BtnAsyncNet, instant);
 	m_btnMgr.hide(m_config3LblOcarina, instant);
 	m_btnMgr.hide(m_config3BtnOcarina, instant);
-	for (u32 i = 0; i < ARRAY_SIZE(m_config3LblUser); ++i)
-		if (m_config3LblUser[i] != -1u)
+	for(u8 i = 0; i < ARRAY_SIZE(m_config3LblUser); ++i)
+		if(m_config3LblUser[i] != (u16)-1)
 			m_btnMgr.hide(m_config3LblUser[i], instant);
 }
 
@@ -77,8 +77,8 @@ void CMenu::_showConfig3(void)
 	m_btnMgr.show(m_config3LblOcarina);
 	m_btnMgr.show(m_config3BtnOcarina);
 
-	for (u32 i = 0; i < ARRAY_SIZE(m_config3LblUser); ++i)
-		if (m_config3LblUser[i] != -1u)
+	for(u8 i = 0; i < ARRAY_SIZE(m_config3LblUser); ++i)
+		if(m_config3LblUser[i] != (u16)-1)
 			m_btnMgr.show(m_config3LblUser[i]);
 
 	int i;
